split sdk
