+++ packages/runner-sdk/lib/index.ts
@@ -1,3 +1,4 @@
+export * from './action.js';
 export * from './dataValidation.js';
 export * from './errors.js';
-export * from './sdk.js';
+export * from './sync.js';
